Moving data fetching back to wrapper component so I can utilize the spinner
